+++ cnccode/data_classes/DBEUser.inc.php
@@ -73,6 +73,7 @@ class DBEUser extends DBEntity
     const streamOneLicenseManagement             = "streamOneLicenseManagement";
     const excludeFromSDManagerDashboard          = "excludeFromSDManagerDashboard";
     const holdAllSRsforQAReview                  = "holdAllSRsforQAReview";
+    const bccOnCustomerEmails                    = "bccOnCustomerEmails";
 
     /**
      * calls constructor()
@@ -408,7 +409,8 @@ class DBEUser extends DBEntity
         $this->addColumn(self::globalExpenseApprover, DA_BOOLEAN, DA_NOT_NULL, null, 0);
         $this->addColumn(self::additionalTimeLevelApprover, DA_BOOLEAN, DA_NOT_NULL, null, 0);
         $this->addColumn(self::sendEmailWhenAssignedService, DA_BOOLEAN, DA_NOT_NULL, null, 1);
-        $this->addColumn(self::holdAllSRsforQAReview, DA_BOOLEAN, DA_NOT_NULL);
+        $this->addColumn(self::holdAllSRsforQAReview, DA_BOOLEAN, DA_NOT_NULL, null, 0);
+        $this->addColumn(self::bccOnCustomerEmails, DA_BOOLEAN, DA_NOT_NULL, null, 0);
         $this->setPK(0);
         $this->setAddColumnsOff();
     }