+++ htdocs/KPIReport.php
@@ -1,0 +1,24 @@
+<?php
+require_once("config.inc.php");
+require_once($cfg["path_ct"]."/CTKPIReport.inc.php");
+session_start();
+page_open(
+	array(
+		'sess' => PHPLIB_CLASSNAME_SESSION,
+		'auth' => PHPLIB_CLASSNAME_AUTH,
+		'perm' => PHPLIB_CLASSNAME_PERM,
+		''
+	)
+);
+GLOBAL $cfg;
+header("Cache-control: private");
+$ctKPIReport= new CTKPIReport(
+	$_SERVER['REQUEST_METHOD'],
+	$_POST,
+	$_GET,
+	$_COOKIE,
+	$cfg
+);
+$ctKPIReport->execute();
+page_close();
+?>