+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 This project changes will be shown here.
 
+## [v5.2.0] - 2021-01-12
+### Added
+- Issue 1038 - Add means of monitoring the licensing data received from StreamOne.
+- Issue 1160 - KPI Graphs for performance
+### Changed
+- Issue 1170 - Modal Popup GUI Improvements
+- Issue 1164 - Combine Dashboards for Change, Sales and Time Requests
+- Issue 1148 - Set max number of mistakes in editor
+- Issue 1179 - Change Summary Of Resolution min character required.
+
 ## [v5.1.4] - 2021-01-05
 ### Fixed
 - Fixed issue with holdForQA accepting null value.